Fix duplicated objects
