+++ modules/effects/arrow-effects-reactor/src/main/kotlin/arrow/effects/CoroutineContextReactorScheduler.kt
@@ -1,28 +1,30 @@
 package arrow.effects
 
-import io.reactivex.Scheduler
-import io.reactivex.disposables.Disposable
-import io.reactivex.internal.disposables.EmptyDisposable
-import java.util.concurrent.TimeUnit
+import reactor.core.Disposable
+import reactor.core.scheduler.Scheduler
 import kotlin.coroutines.experimental.Continuation
 import kotlin.coroutines.experimental.CoroutineContext
 import kotlin.coroutines.experimental.startCoroutine
 
-object CoroutineContextScheduler {
+object CoroutineContextReactorScheduler {
   private interface NonCancellableContinuation : Continuation<Unit>, Disposable
 
+  private val emptyDisposable = object : Disposable {
+    override fun dispose() {}
+
+    override fun isDisposed(): Boolean = true
+  }
+
   fun CoroutineContext.asScheduler(): Scheduler =
-    object : Scheduler() {
-      override fun createWorker(): Worker =
-        object : Worker() {
-          @Volatile
-          var once = false
+    object : Scheduler {
+      override fun schedule(task: Runnable): Disposable =
+        createWorker().schedule(task)
 
-          override fun isDisposed(): Boolean = once
-
-          override fun schedule(run: Runnable, delay: Long, unit: TimeUnit): Disposable {
+      override fun createWorker(): Scheduler.Worker =
+        object : Scheduler.Worker {
+          override fun schedule(run: Runnable): Disposable {
             if (once) {
-              return EmptyDisposable.INSTANCE
+              return emptyDisposable
             }
 
             val a: suspend () -> Unit = { run.run() }
@@ -30,6 +32,11 @@ object CoroutineContextScheduler {
             a.startCoroutine(completion)
             return completion
           }
+
+          @Volatile
+          var once = false
+
+          override fun isDisposed(): Boolean = once
 
           override fun dispose() {
             once = false
