+++ modules/effects/arrow-effects-rx2/src/main/kotlin/arrow/effects/CoroutineContextRx2Scheduler.kt
@@ -1,30 +1,28 @@
 package arrow.effects
 
-import reactor.core.Disposable
-import reactor.core.scheduler.Scheduler
+import io.reactivex.Scheduler
+import io.reactivex.disposables.Disposable
+import io.reactivex.internal.disposables.EmptyDisposable
+import java.util.concurrent.TimeUnit
 import kotlin.coroutines.experimental.Continuation
 import kotlin.coroutines.experimental.CoroutineContext
 import kotlin.coroutines.experimental.startCoroutine
 
-object CoroutineContextScheduler {
+object CoroutineContextRx2Scheduler {
   private interface NonCancellableContinuation : Continuation<Unit>, Disposable
 
-  private val emptyDisposable = object : Disposable {
-    override fun dispose() {}
-
-    override fun isDisposed(): Boolean = true
-  }
-
   fun CoroutineContext.asScheduler(): Scheduler =
-    object : Scheduler {
-      override fun schedule(task: Runnable): Disposable =
-        createWorker().schedule(task)
+    object : Scheduler() {
+      override fun createWorker(): Worker =
+        object : Worker() {
+          @Volatile
+          var once = false
 
-      override fun createWorker(): Scheduler.Worker =
-        object : Scheduler.Worker {
-          override fun schedule(run: Runnable): Disposable {
+          override fun isDisposed(): Boolean = once
+
+          override fun schedule(run: Runnable, delay: Long, unit: TimeUnit): Disposable {
             if (once) {
-              return emptyDisposable
+              return EmptyDisposable.INSTANCE
             }
 
             val a: suspend () -> Unit = { run.run() }
@@ -32,11 +30,6 @@ object CoroutineContextScheduler {
             a.startCoroutine(completion)
             return completion
           }
-
-          @Volatile
-          var once = false
-
-          override fun isDisposed(): Boolean = once
 
           override fun dispose() {
             once = false
