+++ modules/effects/arrow-effects-rx2/src/main/kotlin/arrow/effects/FlowableK.kt
@@ -2,7 +2,7 @@ package arrow.effects
 
 import arrow.Kind
 import arrow.core.*
-import arrow.effects.CoroutineContextScheduler.asScheduler
+import arrow.effects.CoroutineContextRx2Scheduler.asScheduler
 import arrow.effects.typeclasses.Proc
 import arrow.higherkind
 import arrow.typeclasses.Applicative
@@ -44,11 +44,10 @@ data class FlowableK<A>(val flowable: Flowable<A>) : FlowableKOf<A>, FlowableKKi
     return Eval.defer { loop(this) }
   }
 
-  fun <G, B> traverse(GA: Applicative<G>, f: (A) -> Kind<G, B>): Kind<G, FlowableK<B>> = GA.run {
-    foldRight(Eval.always { just(Flowable.empty<B>().k()) }) { a, eval ->
-      f(a).map2Eval(eval) { Flowable.concat(Flowable.just<B>(it.a), it.b.flowable).k() }
+  fun <G, B> traverse(GA: Applicative<G>, f: (A) -> Kind<G, B>): Kind<G, FlowableK<B>> =
+    foldRight(Eval.always { GA.just(Flowable.empty<B>().k()) }) { a, eval ->
+      GA.run { f(a).map2Eval(eval) { Flowable.concat(Flowable.just<B>(it.a), it.b.flowable).k() } }
     }.value()
-  }
 
   fun handleErrorWith(function: (Throwable) -> FlowableK<A>): FlowableK<A> =
     flowable.onErrorResumeNext { t: Throwable -> function(t).flowable }.k()
